+++ packages/base/src/format.js
@@ -12,6 +12,6 @@ module.exports = {
     node: true
   },
 
-  extends: ['./configs/json', './configs/lint', './configs/format']
+  extends: ['./configs/json', './configs/format']
 
 }