+++ packages/typescript/src/configs/format.js
@@ -1,7 +1,6 @@
 module.exports = Object.freeze({
-  extends: ["@fescherer/eslint-config-base:format"],
   rules: Object.freeze({
   // '@typescript-eslint/comma-spacing': ['error', { before: false, after: true }]
-  '@typescript-eslint/member-delimiter-style': ['error', { multiline: { delimiter: 'none' } }]
+    '@typescript-eslint/member-delimiter-style': ['error', { multiline: { delimiter: 'none' } }]
   })
 })