+++ packages/typescript/src/configs/rules-all.js
@@ -1,6 +1,0 @@
-module.exports = Object.freeze({
-  extends: [
-    './rules-format',
-    './rules-lint'
-  ]
-})