+++ packages/typescript/src/format.js
@@ -11,6 +11,7 @@ module.exports = {
 
   parser: '@typescript-eslint/parser',
   plugins: ['@typescript-eslint'],
-  extends: ['./configs/lint', './configs/format']
+
+  extends: ['./configs/format']
 
 };