+++ packages/typescript/src/lint.js
@@ -11,6 +11,6 @@ module.exports = {
 
   parser: '@typescript-eslint/parser',
   plugins: ['@typescript-eslint'],
-  extends: ['./configs/lint', './configs/format']
+  extends: ['./configs/lint']
 
 };