+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @fescherer/eslint-config-react
 
+## 1.3.1
+
+### Patch Changes
+
+- Publish fixes
+- Updated dependencies
+  - @fescherer/eslint-config-typescript@1.3.1
+
 ## 1.3.0
 
 ### Minor Changes