+++ packages/typescript/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @fescherer/eslint-config-typescript
 
+## 1.3.1
+
+### Patch Changes
+
+- Publish fixes
+- Updated dependencies
+  - @fescherer/eslint-config-base@1.3.5
+
 ## 1.3.0
 
 ### Minor Changes